automatik push

+++ Cpp/H/IDT.h
@@ -49,7 +49,7 @@ private:
 public:
   InitIDT(){
     idtr.Limit = 0x0FFF;
-    idtr.Offset = malloc(4096);
+    idtr.Offset = (uint64_t)malloc(4096);
 
     IDT64* Keyboard = (IDT64*)(idtr.Offset + 0x21 * sizeof(IDT64));
     Keyboard->Set_Offset((uint64_t)isr1_handler);
